refactor: performance, thread-safety, DST-safe date math, and int64 IDs

+++ internal/fsapi/tickets.go
@@ -19,13 +19,13 @@ func (c *Client) ListTickets(ctx context.Context, query url.Values) ([]Ticket, b
 
 // LatestConversation returns the most recent conversation for a ticket, or
 // nil when the ticket has none. Conversations are latest-first per the API.
-func (c *Client) LatestConversation(ctx context.Context, ticketID float64) (*Conversation, error) {
+func (c *Client) LatestConversation(ctx context.Context, ticketID int64) (*Conversation, error) {
 	q := url.Values{
 		"per_page":   {"1"},
 		"order_by":   {"created_at"},
 		"order_type": {"desc"},
 	}
-	data, err := c.Get(ctx, fmt.Sprintf("tickets/%.0f/conversations", ticketID), q)
+	data, err := c.Get(ctx, fmt.Sprintf("tickets/%d/conversations", ticketID), q)
 	if err != nil {
 		return nil, err
 	}
@@ -40,11 +40,11 @@ func (c *Client) LatestConversation(ctx context.Context, ticketID float64) (*Con
 }
 
 // UpdateTicket sends a PUT to update ticket fields.
-func (c *Client) UpdateTicket(ctx context.Context, ticketID float64, body any) error {
+func (c *Client) UpdateTicket(ctx context.Context, ticketID int64, body any) error {
 	payload, err := json.Marshal(body)
 	if err != nil {
 		return err
 	}
-	_, err = c.Put(ctx, fmt.Sprintf("tickets/%.0f", ticketID), payload)
+	_, err = c.Put(ctx, fmt.Sprintf("tickets/%d", ticketID), payload)
 	return err
 }
